fix: 커뮤니티 - 좋아요 방어 코딩 수정, 리뷰 작성 1~100 글자 방어

+++ src/main/java/com/placeHere/server/controller/community/CommunityRestController.java
@@ -8,6 +8,7 @@ import com.placeHere.server.domain.User;
 import com.placeHere.server.service.community.CommunityService;
 import com.placeHere.server.service.community.FriendService;
 import com.placeHere.server.service.like.LikeService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.RequestEntity;
@@ -35,31 +36,46 @@ public class CommunityRestController {
 
     //좋아요 추가
     @PostMapping("/addLike")
-    public ResponseEntity<String> addLike(@RequestBody Like like) {
+    public ResponseEntity<Integer> addLike(HttpSession session, @RequestBody Like like) {
 
-        try {
-            Like chkLike = likeService.chkLike(like);
-            System.out.println("/addLike 좋아요체크 : "+chkLike);
-            // 현재 사용자의 좋아요 상태 확인
-            if (chkLike != null) {
+        User user = (User) session.getAttribute("user");
 
-                // 좋아요가 존재하면 취소
-                likeService.removeLike(chkLike);
-                System.out.println("removeLike:: "+  likeService.removeLike(chkLike));
-                //return ResponseEntity.ok("좋아요가 취소 되었습니다");
-                return ResponseEntity.ok("-1");
+        if (user == null) {
+            return ResponseEntity.ok(0);
+        }
 
-            } else {
+        else {
 
-                //좋아요가 없으면 추가
-                likeService.addLike(like.getUserName(), like.getRelationNo(), like.getTarget());
-                //return ResponseEntity.ok("좋아요가 추가되었습니다.");
-                return ResponseEntity.ok("1");
+            try {
 
+                like.setUserName(user.getUsername());
+
+                Like chkLike = likeService.chkLike(like);
+                System.out.println("/addLike 좋아요체크 : " + chkLike);
+
+                // 현재 사용자의 좋아요 상태 확인
+                if (chkLike != null) {
+
+                    // 좋아요가 존재하면 취소
+                    likeService.removeLike(chkLike);
+                    System.out.println("removeLike:: " + likeService.removeLike(chkLike));
+                    //return ResponseEntity.ok("좋아요가 취소 되었습니다");
+                    return ResponseEntity.ok(-1);
+
+                } else {
+
+                    //좋아요가 없으면 추가
+                    likeService.addLike(like.getUserName(), like.getRelationNo(), like.getTarget());
+                    //return ResponseEntity.ok("좋아요가 추가되었습니다.");
+                    return ResponseEntity.ok(1);
+
+                }
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return ResponseEntity.status(500).body(-2);
             }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseEntity.status(500).body("좋아요 오류 ");
+
         }
 
     }
